[refactor/#14]: subject 대신 timer 로 리팩토링
