+++ src/main/java/com/woohaengshi/backend/service/timer/TimerService.java
@@ -1,7 +1,7 @@
-package com.woohaengshi.backend.service.subject;
+package com.woohaengshi.backend.service.timer;
 
 import com.woohaengshi.backend.dto.response.studyrecord.ShowTimerResponse;
 
-public interface SubjectService {
+public interface TimerService {
     ShowTimerResponse getTimer(Long memberId);
 }
