+++ src/test/java/com/woohaengshi/backend/controller/TimerControllerTest.java
@@ -9,7 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class SubjectControllerTest {
+class TimerControllerTest {
     @LocalServerPort private int port;
 
     @BeforeEach
